chore: refactor folders
