+++ PagarMe.Mpos/Callbacks/MposFinishTransactionCallback.cs
@@ -1,14 +1,15 @@
-using PagarMe.Mpos.Helpers;
 using System.Threading.Tasks;
-using static PagarMe.Mpos.Mpos;
+using PagarMe.Mpos.Entities;
+using PagarMe.Mpos.Natives;
+using static PagarMe.Mpos.Natives.Native;
 
 namespace PagarMe.Mpos.Callbacks
 {
     class MposFinishTransactionCallback
     {
-        public static Native.MposFinishTransactionCallbackDelegate Callback(Mpos mpos, TaskCompletionSource<bool> source)
+        public static MposFinishTransactionCallbackDelegate Callback(Mpos mpos, TaskCompletionSource<bool> source)
         {
-            return GCHelper.ManualFree<Native.MposFinishTransactionCallbackDelegate>(releaseGC =>
+            return GCHelper.ManualFree<MposFinishTransactionCallbackDelegate>(releaseGC =>
             {
                 return (mposPtr, err) =>
                 {
@@ -18,12 +19,12 @@ namespace PagarMe.Mpos.Callbacks
             });
         }
 
-        private static Native.Error callback(Mpos mpos, TaskCompletionSource<bool> source, int err)
+        private static Error callback(Mpos mpos, TaskCompletionSource<bool> source, int err)
         {
             mpos.OnFinishedTransaction(err);
             source.SetResult(true);
 
-            return Native.Error.Ok;
+            return Error.Ok;
         }
     }
 }
