+++ PagarMe.Mpos/Callbacks/MposGetTableVersionToInterop.cs
@@ -1,5 +1,5 @@
-using PagarMe.Mpos.Helpers;
-using static PagarMe.Mpos.Mpos.Native;
+using PagarMe.Mpos.Natives;
+using static PagarMe.Mpos.Natives.Native;
 
 namespace PagarMe.Mpos.Callbacks
 {
