+++ PagarMe.Mpos/Callbacks/MposTablesLoadedPaymentCallback.cs
@@ -1,15 +1,16 @@
-using PagarMe.Mpos.Helpers;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using static PagarMe.Mpos.Mpos;
+using PagarMe.Mpos.Entities;
+using PagarMe.Mpos.Natives;
+using static PagarMe.Mpos.Natives.Native;
 
 namespace PagarMe.Mpos.Callbacks
 {
     class MposTablesLoadedPaymentCallback
     {
-        public static Native.MposTablesLoadedCallbackDelegate Callback(Mpos mpos, int amount, IEnumerable<EmvApplication> applications, PaymentMethod magstripePaymentMethod, TaskCompletionSource<PaymentResult> source)
+        public static MposTablesLoadedCallbackDelegate Callback(Mpos mpos, int amount, IEnumerable<EmvApplication> applications, PaymentMethod magstripePaymentMethod, TaskCompletionSource<PaymentResult> source)
         {
-            return GCHelper.ManualFree<Native.MposTablesLoadedCallbackDelegate>(releaseGC =>
+            return GCHelper.ManualFree<MposTablesLoadedCallbackDelegate>(releaseGC =>
             {
                 return (mposPtr, tableError, loaded) =>
                 {
@@ -20,37 +21,37 @@ namespace PagarMe.Mpos.Callbacks
             });
         }
 
-        private static Native.Error callback(Mpos mpos, int amount, IEnumerable<EmvApplication> applications, PaymentMethod magstripePaymentMethod, TaskCompletionSource<PaymentResult> source)
+        private static Error callback(Mpos mpos, int amount, IEnumerable<EmvApplication> applications, PaymentMethod magstripePaymentMethod, TaskCompletionSource<PaymentResult> source)
         {
             var callback = MposPaymentCallback.Callback(mpos, source);
 
-            var acquirers = new List<Native.Acquirer>();
-            var riskProfiles = new List<Native.RiskManagement>();
+            var acquirers = new List<Acquirer>();
+            var riskProfiles = new List<RiskManagement>();
 
-            var rawApplications = new List<Native.Application>();
+            var rawApplications = new List<Application>();
             if (applications != null)
                 foreach (var application in applications)
                 {
                     var entry = mpos.TMSStorage.SelectApplication(application.Brand, (int)application.PaymentMethod);
-                    if (entry != null) rawApplications.Add(new Native.Application(entry));
+                    if (entry != null) rawApplications.Add(new Application(entry));
                 }
             else
                 foreach (var entry in mpos.TMSStorage.GetApplicationRows())
-                    rawApplications.Add(new Native.Application(entry));
+                    rawApplications.Add(new Application(entry));
 
-            foreach (var entry in mpos.TMSStorage.GetAcquirerRows()) acquirers.Add(new Native.Acquirer(entry));
+            foreach (var entry in mpos.TMSStorage.GetAcquirerRows()) acquirers.Add(new Acquirer(entry));
 
             foreach (var entry in mpos.TMSStorage.GetRiskManagementRows())
-                riskProfiles.Add(new Native.RiskManagement(entry));
+                riskProfiles.Add(new RiskManagement(entry));
 
-            var error = Native.ProcessPayment(mpos.nativeMpos, amount, rawApplications.ToArray(), rawApplications.Count,
+            var error = ProcessPayment(mpos.nativeMpos, amount, rawApplications.ToArray(), rawApplications.Count,
                 acquirers.ToArray(), acquirers.Count, riskProfiles.ToArray(), riskProfiles.Count,
                 (int)magstripePaymentMethod, callback);
 
-            if (error != Native.Error.Ok)
+            if (error != Error.Ok)
                 throw new MposException(error);
 
-            return Native.Error.Ok;
+            return Error.Ok;
         }
     }
 }
