+++ PagarMe.Mpos/Callbacks/TmsStoreToInterop.cs
@@ -1,6 +1,7 @@
-using PagarMe.Mpos.Helpers;
 using System.Collections.Generic;
-using static PagarMe.Mpos.Mpos.Native;
+using PagarMe.Mpos.Entities;
+using PagarMe.Mpos.Natives;
+using static PagarMe.Mpos.Natives.Native;
 
 namespace PagarMe.Mpos.Callbacks
 {
