+++ PagarMe.Mpos/Entities/Decision.cs
@@ -1,0 +1,9 @@
+namespace PagarMe.Mpos.Entities
+{
+	public enum Decision
+	{
+		Approved = 0,
+		Refused,
+		GoOnline
+	}
+}
