+++ PagarMe.Mpos/Entities/EmvApplication.cs
@@ -1,4 +1,4 @@
-namespace PagarMe.Mpos
+namespace PagarMe.Mpos.Entities
 {
     public struct EmvApplication
     {
