+++ PagarMe.Mpos/Entities/Error.cs
@@ -1,0 +1,9 @@
+namespace PagarMe.Mpos.Entities
+{
+	internal enum Error
+	{
+		ConnError = -1,
+		Ok = 0,
+		Error
+	}
+}
