+++ PagarMe.Mpos/Entities/TransactionStatus.cs
@@ -1,0 +1,9 @@
+namespace PagarMe.Mpos.Entities
+{
+	public enum TransactionStatus
+	{
+		Ok = 0,
+		Error = 1,
+		NonZero = 9
+	}
+}
