+++ PagarMe.Mpos/MposException.cs
@@ -1,10 +1,11 @@
 using System;
+using PagarMe.Mpos.Entities;
 
 namespace PagarMe.Mpos
 {
     public class MposException : Exception
     {
-        internal MposException(Mpos.Native.Error error)
+        internal MposException(Error error)
             : base("An error ocurred: " + error)
         {
         }
