+++ PagarMe.Mpos/PagarMe.Mpos.csproj
@@ -62,28 +62,37 @@
     <Compile Include="Callbacks\MposTablesLoadedSynchronizeTablesCallback.cs" />
     <Compile Include="Callbacks\TmsStoreCallback.cs" />
     <Compile Include="Callbacks\TmsStoreToInterop.cs" />
-    <Compile Include="Helpers\GCHelper.cs" />
-    <Compile Include="Helpers\PtrHelper.cs" />
-    <Compile Include="INativeImport.cs" />
-    <Compile Include="Mpos.NativeBit32.cs" />
-    <Compile Include="Mpos.NativeBit64.cs" />
+    <Compile Include="Natives\GCHelper.cs" />
+    <Compile Include="Natives\PtrHelper.cs" />
+    <Compile Include="Natives\INativeImport.cs" />
+    <Compile Include="Entities\Acquirer.cs" />
+    <Compile Include="Entities\Aid.cs" />
+    <Compile Include="Entities\Application.cs" />
+    <Compile Include="Entities\Capk.cs" />
+    <Compile Include="Entities\Decision.cs" />
+    <Compile Include="Entities\Error.cs" />
+    <Compile Include="Natives\NativeBit32.cs" />
+    <Compile Include="Natives\NativeBit64.cs" />
     <Compile Include="Callbacks\MposPaymentCallback.cs" />
-    <Compile Include="Mpos.NativeConverter.cs" />
+    <Compile Include="Natives\NativeConverter.cs" />
     <Compile Include="Mpos.cs" />
-    <Compile Include="PaymentResult.cs" />
-    <Compile Include="PaymentStatus.cs" />
+    <Compile Include="Entities\PaymentInfo.cs" />
+    <Compile Include="Entities\RiskManagement.cs" />
+    <Compile Include="Entities\TransactionStatus.cs" />
+    <Compile Include="Entities\PaymentResult.cs" />
+    <Compile Include="Entities\PaymentStatus.cs" />
     <Compile Include="MposException.cs" />
-    <Compile Include="PaymentMethod.cs" />
-    <Compile Include="ApiHelper.cs" />
-    <Compile Include="ByteArrayConverter.cs" />
-    <Compile Include="CaptureMethod.cs" />
-    <Compile Include="EmvApplication.cs" />
+    <Compile Include="Entities\PaymentMethod.cs" />
+    <Compile Include="Api\ApiHelper.cs" />
+    <Compile Include="Natives\ByteArrayConverter.cs" />
+    <Compile Include="Entities\CaptureMethod.cs" />
+    <Compile Include="Entities\EmvApplication.cs" />
     <Compile Include="Devices\DeviceKind.cs" />
     <Compile Include="Devices\DeviceManager.cs" />
     <Compile Include="Devices\IDevice.cs" />
     <Compile Include="Devices\SerialDevice.cs" />
     <Compile Include="Properties\MposAssemblyInfo.cs" />
-    <Compile Include="TLS.cs" />
+    <Compile Include="Api\TLS.cs" />
     <Compile Include="Tms\Bit32.cs" />
     <Compile Include="Tms\Linux.cs" />
     <Compile Include="Tms\Bit64.cs" />
@@ -92,7 +101,7 @@
     <Compile Include="Tms\SQLite.cs" />
     <Compile Include="Tms\SQLiteAsync.cs" />
     <Compile Include="Abecs\AbecsStream.cs" />
-    <Compile Include="Mpos.Native.cs" />
+    <Compile Include="Natives\Native.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
