+++ PagarMe.Mpos/Tms/Linux.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Runtime.InteropServices;
-using static SQLite.SQLite3;
 
-namespace SQLite
+namespace PagarMe.Mpos.Tms
 {
     class Linux : ISqliteImport
     {
@@ -20,80 +19,80 @@ namespace SQLite
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_open", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result OpenInternal([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db);
-        public Result Open([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db)
+        public static extern SQLite3.Result OpenInternal([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db);
+        public SQLite3.Result Open([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db)
         {
             return OpenInternal(filename, out db);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_open_v2", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result OpenInternal([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db, int flags, IntPtr zvfs);
-        public Result Open([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db, int flags, IntPtr zvfs)
+        public static extern SQLite3.Result OpenInternal([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db, int flags, IntPtr zvfs);
+        public SQLite3.Result Open([MarshalAs(UnmanagedType.LPStr)] string filename, out IntPtr db, int flags, IntPtr zvfs)
         {
             return OpenInternal(filename, out db, flags, zvfs);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_open_v2", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result OpenInternal(byte[] filename, out IntPtr db, int flags, IntPtr zvfs);
-        public Result Open(byte[] filename, out IntPtr db, int flags, IntPtr zvfs)
+        public static extern SQLite3.Result OpenInternal(byte[] filename, out IntPtr db, int flags, IntPtr zvfs);
+        public SQLite3.Result Open(byte[] filename, out IntPtr db, int flags, IntPtr zvfs)
         {
             return OpenInternal(filename, out db, flags, zvfs);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_open16", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result Open16Internal([MarshalAs(UnmanagedType.LPWStr)] string filename, out IntPtr db);
-        public Result Open16([MarshalAs(UnmanagedType.LPWStr)] string filename, out IntPtr db)
+        public static extern SQLite3.Result Open16Internal([MarshalAs(UnmanagedType.LPWStr)] string filename, out IntPtr db);
+        public SQLite3.Result Open16([MarshalAs(UnmanagedType.LPWStr)] string filename, out IntPtr db)
         {
             return Open16Internal(filename, out db);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_enable_load_extension", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result EnableLoadExtensionInternal(IntPtr db, int onoff);
-        public Result EnableLoadExtension(IntPtr db, int onoff)
+        public static extern SQLite3.Result EnableLoadExtensionInternal(IntPtr db, int onoff);
+        public SQLite3.Result EnableLoadExtension(IntPtr db, int onoff)
         {
             return EnableLoadExtensionInternal(db, onoff);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_close", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result CloseInternal(IntPtr db);
-        public Result Close(IntPtr db)
+        public static extern SQLite3.Result CloseInternal(IntPtr db);
+        public SQLite3.Result Close(IntPtr db)
         {
             return CloseInternal(db);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_close_v2", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result Close2Internal(IntPtr db);
-        public Result Close2(IntPtr db)
+        public static extern SQLite3.Result Close2Internal(IntPtr db);
+        public SQLite3.Result Close2(IntPtr db)
         {
             return Close2Internal(db);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_initialize", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result InitializeInternal();
-        public Result Initialize()
+        public static extern SQLite3.Result InitializeInternal();
+        public SQLite3.Result Initialize()
         {
             return InitializeInternal();
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_shutdown", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result ShutdownInternal();
-        public Result Shutdown()
+        public static extern SQLite3.Result ShutdownInternal();
+        public SQLite3.Result Shutdown()
         {
             return ShutdownInternal();
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_config", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result ConfigInternal(ConfigOption option);
-        public Result Config(ConfigOption option)
+        public static extern SQLite3.Result ConfigInternal(SQLite3.ConfigOption option);
+        public SQLite3.Result Config(SQLite3.ConfigOption option)
         {
             return ConfigInternal(option);
         }
@@ -108,8 +107,8 @@ namespace SQLite
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_busy_timeout", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result BusyTimeoutInternal(IntPtr db, int milliseconds);
-        public Result BusyTimeout(IntPtr db, int milliseconds)
+        public static extern SQLite3.Result BusyTimeoutInternal(IntPtr db, int milliseconds);
+        public SQLite3.Result BusyTimeout(IntPtr db, int milliseconds)
         {
             return BusyTimeoutInternal(db, milliseconds);
         }
@@ -124,8 +123,8 @@ namespace SQLite
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_prepare_v2", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result Prepare2Internal(IntPtr db, [MarshalAs(UnmanagedType.LPStr)] string sql, int numBytes, out IntPtr stmt, IntPtr pzTail);
-        public Result Prepare2(IntPtr db, [MarshalAs(UnmanagedType.LPStr)] string sql, int numBytes, out IntPtr stmt, IntPtr pzTail)
+        public static extern SQLite3.Result Prepare2Internal(IntPtr db, [MarshalAs(UnmanagedType.LPStr)] string sql, int numBytes, out IntPtr stmt, IntPtr pzTail);
+        public SQLite3.Result Prepare2(IntPtr db, [MarshalAs(UnmanagedType.LPStr)] string sql, int numBytes, out IntPtr stmt, IntPtr pzTail)
         {
             return Prepare2Internal(db, sql, numBytes, out stmt, pzTail);
         }
@@ -141,24 +140,24 @@ namespace SQLite
 #endif
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_step", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result StepInternal(IntPtr stmt);
-        public Result Step(IntPtr stmt)
+        public static extern SQLite3.Result StepInternal(IntPtr stmt);
+        public SQLite3.Result Step(IntPtr stmt)
         {
             return StepInternal(stmt);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_reset", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result ResetInternal(IntPtr stmt);
-        public Result Reset(IntPtr stmt)
+        public static extern SQLite3.Result ResetInternal(IntPtr stmt);
+        public SQLite3.Result Reset(IntPtr stmt)
         {
             return ResetInternal(stmt);
         }
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_finalize", CallingConvention = CallingConvention.Cdecl)]
-        public static extern Result FinalizeInternal(IntPtr stmt);
-        public Result Finalize(IntPtr stmt)
+        public static extern SQLite3.Result FinalizeInternal(IntPtr stmt);
+        public SQLite3.Result Finalize(IntPtr stmt)
         {
             return FinalizeInternal(stmt);
         }
@@ -261,8 +260,8 @@ namespace SQLite
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_column_type", CallingConvention = CallingConvention.Cdecl)]
-        public static extern ColType ColumnTypeInternal(IntPtr stmt, int index);
-        public ColType ColumnType(IntPtr stmt, int index)
+        public static extern SQLite3.ColType ColumnTypeInternal(IntPtr stmt, int index);
+        public SQLite3.ColType ColumnType(IntPtr stmt, int index)
         {
             return ColumnTypeInternal(stmt, index);
         }
@@ -325,8 +324,8 @@ namespace SQLite
 
 
         [DllImport(LibraryPath, EntryPoint = "sqlite3_extended_errcode", CallingConvention = CallingConvention.Cdecl)]
-        public static extern ExtendedResult ExtendedErrCodeInternal(IntPtr db);
-        public ExtendedResult ExtendedErrCode(IntPtr db)
+        public static extern SQLite3.ExtendedResult ExtendedErrCodeInternal(IntPtr db);
+        public SQLite3.ExtendedResult ExtendedErrCode(IntPtr db)
         {
             return ExtendedErrCodeInternal(db);
         }
