+++ PagarMe.Mpos/Tms/SQLiteAsync.cs
@@ -30,7 +30,7 @@ using System.Threading.Tasks;
 
 #pragma warning disable 1591 // XML Doc Comments
 
-namespace SQLite
+namespace PagarMe.Mpos.Tms
 {
 	public partial class SQLiteAsyncConnection
 	{
@@ -248,7 +248,7 @@ namespace SQLite
 		public Task RunInTransactionAsync (Action<SQLiteAsyncConnection> action)
 		{
 			return Task.Factory.StartNew (() => {
-				var conn = this.GetConnection ();
+				var conn = GetConnection ();
 				using (conn.Lock ()) {
 					conn.BeginTransaction ();
 					try {
@@ -267,7 +267,7 @@ namespace SQLite
 		{
 			return Task.Factory.StartNew(() =>
 				{
-					var conn = this.GetConnection();
+					var conn = GetConnection();
 					using (conn.Lock())
 					{
 						conn.BeginTransaction();
@@ -410,7 +410,7 @@ namespace SQLite
 
 		internal CreateTablesResult ()
 		{
-			this.Results = new Dictionary<Type, int> ();
+			Results = new Dictionary<Type, int> ();
 		}
 	}
